add basketball data, change teams rendering for different sports
