+++ app/assets/stylesheets/views/fantasy_teams/transfers/index.scss
@@ -1,11 +1,5 @@
 @import '../../../general/colors.scss';
 
-.position-members {
-  display: flex;
-  flex-direction: row;
-  margin-right: 1rem;
-}
-
 #fantasy-team-transfers {
   display: flex;
   flex-direction: row;
