+++ config/sport_settings.json
@@ -6,7 +6,8 @@
       "max_active_players": 11,
       "max_team_players": 3,
       "free_transfers_per_week": 1,
-      "points_per_transfer": 4
+      "points_per_transfer": 4,
+      "changes": true
     },
     "basketball": {
       "name": { "en": "Basketball", "ru": "Баскетбол" },
@@ -14,7 +15,8 @@
       "max_active_players": 10,
       "max_team_players": 2,
       "free_transfers_per_week": 1,
-      "points_per_transfer": 4
+      "points_per_transfer": 4,
+      "changes": false
     },
     "hockey": {
       "name": { "en": "Hockey", "ru": "Хоккей" },
@@ -22,7 +24,8 @@
       "max_active_players": 17,
       "max_team_players": 2,
       "free_transfers_per_week": 1,
-      "points_per_transfer": 4
+      "points_per_transfer": 4,
+      "changes": true
     }
   },
   "positions": {
