+++ spec/services/lineups/players/update_service_spec.rb
@@ -3,30 +3,28 @@
 describe Lineups::Players::UpdateService, type: :service do
   subject(:service_call) {
     described_class.new(
-      update_football_players_service: update_football_players_service
+      players_validator_service: players_validator_service
     ).call(lineup: lineup, lineups_players_params: lineups_players_params)
   }
 
   let!(:lineup) { create :lineup }
-  let(:update_football_players_service) { double }
+  let(:players_validator_service) { double }
+  let(:players_validator) { double }
   let(:call_result) { double }
-  let(:lineups_players_params) { { id: 1, active: true, change_order: 0 } }
+  let(:validator_result) { ['Error'] }
+  let(:lineups_players_params) { [{ id: 1, active: true, change_order: 0 }] }
 
   before do
-    allow(update_football_players_service).to receive(:call).and_return(call_result)
-    allow(call_result).to receive(:errors).and_return(['Error'])
+    allow(players_validator_service).to receive(:new).and_return(players_validator)
+    allow(players_validator).to receive(:call).and_return(validator_result)
   end
 
   context 'for football' do
     context 'for invalid data' do
-      before do
-        allow(call_result).to receive(:failure?).and_return(true)
-      end
-
       it 'calls football_service' do
         service_call
 
-        expect(update_football_players_service).to have_received(:call)
+        expect(players_validator).to have_received(:call)
       end
 
       it 'and it fails' do
@@ -37,14 +35,12 @@ describe Lineups::Players::UpdateService, type: :service do
     end
 
     context 'for valid data' do
-      before do
-        allow(call_result).to receive(:failure?).and_return(false)
-      end
+      let(:validator_result) { [] }
 
       it 'calls football_service' do
         service_call
 
-        expect(update_football_players_service).to have_received(:call)
+        expect(players_validator).to have_received(:call)
       end
 
       it 'and it succeed' do
